Only run the old-releases on \*buntu machines
git-svn-id: https://svn.code.sf.net/p/freeghost/code/trunk@3617 71f96598-fa45-0410-b640-bcd6f8691b32

+++ lib/ubuntu/functions.sh
@@ -490,15 +490,15 @@ installPackages() {
     fi
     dots "Preparing Package Manager"
     apt-get -yq update >/dev/null 2>&1
-    if [ "$?" != 0 ]; then
+    if [ "$?" != 0 ] && [[ "$linuxReleaseName" == +(*'buntu'*) ]]; then
         cp /etc/apt/sources.list /etc/apt/source.list.original_fog
         sed -i -e 's/archive.ubuntu.com\|security.ubuntu.com/old-releases.ubuntu.com/g' /etc/apt/sources.list
         apt-get -yq update >/dev/null 2>&1
-    fi
-    if [ "$?" != 0 ]; then
-        cp -f /etc/apt/sources.list.original_fog /etc/apt/source.list >/dev/null 2>&1
-        rm -f /etc/apt/sources.list.original >/dev/null 2>&1
-        false
+        if [ "$?" != 0 ]; then
+            cp -f /etc/apt/sources.list.original_fog /etc/apt/source.list >/dev/null 2>&1
+            rm -f /etc/apt/sources.list.original >/dev/null 2>&1
+            false
+        fi
     fi
     errorStat $?
 	if [ "$installlang" = "1" ]; then
